Revert "[Backport 3.1] Added tasks scheduling and management mechanism for hy…" (#172)
This reverts commit fd01bc222623cf7876e49f18e2f4b7f4143affa9.
