+++ CHANGELOG.md
@@ -4,22 +4,10 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
-### Breaking Changes
+### Added
 
-### Features
+### Removed
 
-### Enhancements
-* Added tasks scheduling and management mechanism for hybrid optimizer experiments ([#139](https://github.com/opensearch-project/search-relevance/pull/139))
-
-### Bug Fixes
-
-### Infrastructure
-* Added end to end integration tests for experiments ([#154](https://github.com/opensearch-project/search-relevance/pull/154))
+### Fixed
 
 ### Security
-
-### Documentation
-
-### Maintenance
-
-### Refactoring
