+++ src/main/java/org/opensearch/searchrelevance/transport/experiment/PutExperimentTransportAction.java
@@ -8,6 +8,9 @@
 package org.opensearch.searchrelevance.transport.experiment;
 
 import static org.opensearch.searchrelevance.common.MetricsConstants.PAIRWISE_FIELD_NAME_QUERY_TEXT;
+import static org.opensearch.searchrelevance.experiment.ExperimentOptionsForHybridSearch.EXPERIMENT_OPTION_COMBINATION_TECHNIQUE;
+import static org.opensearch.searchrelevance.experiment.ExperimentOptionsForHybridSearch.EXPERIMENT_OPTION_NORMALIZATION_TECHNIQUE;
+import static org.opensearch.searchrelevance.experiment.ExperimentOptionsForHybridSearch.EXPERIMENT_OPTION_WEIGHTS_FOR_COMBINATION;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -23,23 +26,25 @@ import java.util.stream.Collectors;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.action.index.IndexResponse;
-import org.opensearch.action.search.SearchResponse;
 import org.opensearch.action.support.ActionFilters;
 import org.opensearch.action.support.HandledTransportAction;
+import org.opensearch.cluster.service.ClusterService;
 import org.opensearch.common.inject.Inject;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.rest.RestStatus;
 import org.opensearch.searchrelevance.dao.ExperimentDao;
-import org.opensearch.searchrelevance.dao.JudgmentDao;
+import org.opensearch.searchrelevance.dao.ExperimentVariantDao;
 import org.opensearch.searchrelevance.dao.QuerySetDao;
 import org.opensearch.searchrelevance.dao.SearchConfigurationDao;
 import org.opensearch.searchrelevance.exception.SearchRelevanceException;
-import org.opensearch.searchrelevance.executors.HybridSearchTaskManager;
-import org.opensearch.searchrelevance.experiment.HybridOptimizerExperimentProcessor;
+import org.opensearch.searchrelevance.experiment.ExperimentOptionsFactory;
+import org.opensearch.searchrelevance.experiment.ExperimentOptionsForHybridSearch;
+import org.opensearch.searchrelevance.experiment.ExperimentVariantHybridSearchDTO;
 import org.opensearch.searchrelevance.metrics.MetricsHelper;
 import org.opensearch.searchrelevance.model.AsyncStatus;
 import org.opensearch.searchrelevance.model.Experiment;
 import org.opensearch.searchrelevance.model.ExperimentType;
+import org.opensearch.searchrelevance.model.ExperimentVariant;
 import org.opensearch.searchrelevance.model.QuerySet;
 import org.opensearch.searchrelevance.model.SearchConfiguration;
 import org.opensearch.searchrelevance.utils.TimeUtils;
@@ -51,31 +56,33 @@ import org.opensearch.transport.TransportService;
  */
 public class PutExperimentTransportAction extends HandledTransportAction<PutExperimentRequest, IndexResponse> {
 
+    private final ClusterService clusterService;
     private final ExperimentDao experimentDao;
+    private final ExperimentVariantDao experimentVariantDao;
     private final QuerySetDao querySetDao;
     private final SearchConfigurationDao searchConfigurationDao;
     private final MetricsHelper metricsHelper;
-    private final HybridOptimizerExperimentProcessor hybridOptimizerExperimentProcessor;
 
     private static final Logger LOGGER = LogManager.getLogger(PutExperimentTransportAction.class);
 
     @Inject
     public PutExperimentTransportAction(
+        ClusterService clusterService,
         TransportService transportService,
         ActionFilters actionFilters,
         ExperimentDao experimentDao,
+        ExperimentVariantDao experimentVariantDao,
         QuerySetDao querySetDao,
         SearchConfigurationDao searchConfigurationDao,
-        MetricsHelper metricsHelper,
-        JudgmentDao judgmentDao,
-        HybridSearchTaskManager hybridSearchTaskManager
+        MetricsHelper metricsHelper
     ) {
         super(PutExperimentAction.NAME, transportService, actionFilters, PutExperimentRequest::new);
+        this.clusterService = clusterService;
         this.experimentDao = experimentDao;
+        this.experimentVariantDao = experimentVariantDao;
         this.querySetDao = querySetDao;
         this.searchConfigurationDao = searchConfigurationDao;
         this.metricsHelper = metricsHelper;
-        this.hybridOptimizerExperimentProcessor = new HybridOptimizerExperimentProcessor(judgmentDao, hybridSearchTaskManager);
     }
 
     @Override
@@ -120,100 +127,22 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
     }
 
     private void triggerAsyncProcessing(String experimentId, PutExperimentRequest request) {
-        // First, get QuerySet asynchronously
-        querySetDao.getQuerySet(request.getQuerySetId(), ActionListener.wrap(querySetResponse -> {
-            try {
-                QuerySet querySet = convertToQuerySet(querySetResponse);
-                List<String> queryTextWithReferences = querySet.querySetQueries()
-                    .stream()
-                    .map(e -> e.queryText())
-                    .collect(Collectors.toList());
+        try {
+            QuerySet querySet = querySetDao.getQuerySetSync(request.getQuerySetId());
+            List<String> queryTextWithReferences = querySet.querySetQueries().stream().map(e -> e.queryText()).collect(Collectors.toList());
 
-                // Then get SearchConfigurations asynchronously
-                fetchSearchConfigurationsAsync(experimentId, request, queryTextWithReferences);
-            } catch (Exception e) {
-                handleAsyncFailure(experimentId, request, "Failed to process QuerySet", e);
-            }
-        }, e -> { handleAsyncFailure(experimentId, request, "Failed to fetch QuerySet", e); }));
-    }
-
-    private void fetchSearchConfigurationsAsync(String experimentId, PutExperimentRequest request, List<String> queryTextWithReferences) {
-        Map<String, List<String>> indexAndQueries = new HashMap<>();
-        AtomicInteger pendingConfigs = new AtomicInteger(request.getSearchConfigurationList().size());
-        AtomicBoolean hasFailure = new AtomicBoolean(false);
-
-        for (String configId : request.getSearchConfigurationList()) {
-            searchConfigurationDao.getSearchConfiguration(configId, ActionListener.wrap(searchConfigResponse -> {
-                try {
-                    if (hasFailure.get()) return;
-
-                    SearchConfiguration config = convertToSearchConfiguration(searchConfigResponse);
-                    synchronized (indexAndQueries) {
-                        indexAndQueries.put(config.id(), Arrays.asList(config.index(), config.query(), config.searchPipeline()));
-                    }
-
-                    // Check if all configurations are fetched
-                    if (pendingConfigs.decrementAndGet() == 0) {
-                        calculateMetricsAsync(experimentId, request, indexAndQueries, queryTextWithReferences);
-                    }
-                } catch (Exception e) {
-                    if (hasFailure.compareAndSet(false, true)) {
-                        handleAsyncFailure(experimentId, request, "Failed to process SearchConfiguration", e);
-                    }
-                }
-            }, e -> {
-                if (hasFailure.compareAndSet(false, true)) {
-                    handleAsyncFailure(experimentId, request, "Failed to fetch SearchConfiguration: " + configId, e);
-                }
-            }));
-        }
-    }
-
-    private QuerySet convertToQuerySet(SearchResponse response) {
-        if (response.getHits().getTotalHits().value() == 0) {
-            throw new SearchRelevanceException("QuerySet not found", RestStatus.NOT_FOUND);
-        }
-
-        Map<String, Object> sourceMap = response.getHits().getHits()[0].getSourceAsMap();
-
-        // Convert querySetQueries from list of maps to List<QuerySetEntry>
-        List<org.opensearch.searchrelevance.model.QuerySetEntry> querySetEntries = new ArrayList<>();
-        Object querySetQueriesObj = sourceMap.get("querySetQueries");
-        if (querySetQueriesObj instanceof List) {
-            List<Map<String, Object>> querySetQueriesList = (List<Map<String, Object>>) querySetQueriesObj;
-            querySetEntries = querySetQueriesList.stream()
-                .map(
-                    entryMap -> org.opensearch.searchrelevance.model.QuerySetEntry.Builder.builder()
-                        .queryText((String) entryMap.get("queryText"))
-                        .build()
-                )
+            List<SearchConfiguration> searchConfigurations = request.getSearchConfigurationList()
+                .stream()
+                .map(id -> searchConfigurationDao.getSearchConfigurationSync(id))
                 .collect(Collectors.toList());
+            Map<String, List<String>> indexAndQueries = new HashMap<>();
+            for (SearchConfiguration config : searchConfigurations) {
+                indexAndQueries.put(config.id(), Arrays.asList(config.index(), config.query(), config.searchPipeline()));
+            }
+            calculateMetricsAsync(experimentId, request, indexAndQueries, queryTextWithReferences);
+        } catch (Exception e) {
+            handleAsyncFailure(experimentId, request, "Failed to start async processing", e);
         }
-
-        return org.opensearch.searchrelevance.model.QuerySet.Builder.builder()
-            .id((String) sourceMap.get("id"))
-            .name((String) sourceMap.get("name"))
-            .description((String) sourceMap.get("description"))
-            .timestamp((String) sourceMap.get("timestamp"))
-            .sampling((String) sourceMap.get("sampling"))
-            .querySetQueries(querySetEntries)
-            .build();
-    }
-
-    private SearchConfiguration convertToSearchConfiguration(SearchResponse response) {
-        if (response.getHits().getTotalHits().value() == 0) {
-            throw new SearchRelevanceException("SearchConfiguration not found", RestStatus.NOT_FOUND);
-        }
-
-        Map<String, Object> source = response.getHits().getHits()[0].getSourceAsMap();
-        return new SearchConfiguration(
-            (String) source.get("id"),
-            (String) source.get("name"),
-            (String) source.get("timestamp"),
-            (String) source.get("index"),
-            (String) source.get("query"),
-            (String) source.get("searchPipeline")
-        );
     }
 
     private void calculateMetricsAsync(
@@ -262,10 +191,6 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
         List<String> judgmentList
     ) {
         for (String queryText : queryTexts) {
-            if (hasFailure.get()) {
-                return;
-            }
-
             if (request.getType() == ExperimentType.PAIRWISE_COMPARISON) {
                 metricsHelper.processPairwiseMetrics(
                     queryText,
@@ -286,27 +211,60 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
                     )
                 );
             } else if (request.getType() == ExperimentType.HYBRID_OPTIMIZER) {
-                // Use our task manager implementation for hybrid optimizer
-                hybridOptimizerExperimentProcessor.processHybridOptimizerExperiment(
-                    experimentId,
+                Map<String, Object> defaultParametersForHybridSearch = ExperimentOptionsFactory
+                    .createDefaultExperimentParametersForHybridSearch();
+                ExperimentOptionsForHybridSearch experimentOptionForHybridSearch =
+                    (ExperimentOptionsForHybridSearch) ExperimentOptionsFactory.createExperimentOptions(
+                        ExperimentOptionsFactory.HYBRID_SEARCH_EXPERIMENT_OPTIONS,
+                        defaultParametersForHybridSearch
+                    );
+                List<ExperimentVariantHybridSearchDTO> experimentVariantDTOs = experimentOptionForHybridSearch.getParameterCombinations(
+                    true
+                );
+                List<ExperimentVariant> experimentVariants = new ArrayList<>();
+                for (ExperimentVariantHybridSearchDTO experimentVariantDTO : experimentVariantDTOs) {
+                    Map<String, Object> parameters = new HashMap<>(
+                        Map.of(
+                            EXPERIMENT_OPTION_NORMALIZATION_TECHNIQUE,
+                            experimentVariantDTO.getNormalizationTechnique(),
+                            EXPERIMENT_OPTION_COMBINATION_TECHNIQUE,
+                            experimentVariantDTO.getCombinationTechnique(),
+                            EXPERIMENT_OPTION_WEIGHTS_FOR_COMBINATION,
+                            experimentVariantDTO.getQueryWeightsForCombination()
+                        )
+                    );
+                    String experimentVariantId = UUID.randomUUID().toString();
+                    ExperimentVariant experimentVariant = new ExperimentVariant(
+                        experimentVariantId,
+                        TimeUtils.getTimestamp(),
+                        ExperimentType.HYBRID_OPTIMIZER,
+                        AsyncStatus.PROCESSING,
+                        experimentId,
+                        parameters,
+                        Map.of()
+                    );
+                    experimentVariants.add(experimentVariant);
+                    experimentVariantDao.putExperimentVariant(experimentVariant, ActionListener.wrap(response -> {}, e -> {}));
+                }
+                metricsHelper.processEvaluationMetrics(
                     queryText,
                     indexAndQueries,
-                    judgmentList,
                     request.getSize(),
-                    hasFailure,
-                    ActionListener.wrap(
-                        queryResults -> handleQueryResults(
+                    judgmentList,
+                    ActionListener.wrap(queryResults -> {
+                        Map<String, Object> convertedResults = new HashMap<>(queryResults);
+                        handleQueryResults(
                             queryText,
-                            queryResults,
+                            convertedResults,
                             finalResults,
                             pendingQueries,
                             experimentId,
                             request,
                             hasFailure,
                             judgmentList
-                        ),
-                        error -> handleFailure(error, hasFailure, experimentId, request)
-                    )
+                        );
+                    }, error -> handleFailure(error, hasFailure, experimentId, request)),
+                    experimentVariants
                 );
             } else if (request.getType() == ExperimentType.POINTWISE_EVALUATION) {
                 metricsHelper.processEvaluationMetrics(
@@ -326,8 +284,7 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
                             hasFailure,
                             judgmentList
                         );
-                    }, error -> handleFailure(error, hasFailure, experimentId, request)),
-                    Collections.emptyList()
+                    }, error -> handleFailure(error, hasFailure, experimentId, request))
                 );
             } else {
                 throw new SearchRelevanceException("Unknown experimentType" + request.getType(), RestStatus.BAD_REQUEST);
@@ -349,25 +306,8 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
 
         try {
             synchronized (finalResults) {
-                // Handle different response formats based on experiment type
-                if (request.getType() == ExperimentType.HYBRID_OPTIMIZER) {
-                    // For HYBRID_OPTIMIZER, the response contains searchConfigurationResults
-                    List<Map<String, Object>> searchConfigResults = (List<Map<String, Object>>) queryResults.get(
-                        "searchConfigurationResults"
-                    );
-                    if (searchConfigResults != null) {
-                        for (Map<String, Object> configResult : searchConfigResults) {
-                            Map<String, Object> resultWithQuery = new HashMap<>(configResult);
-                            resultWithQuery.put(PAIRWISE_FIELD_NAME_QUERY_TEXT, queryText);
-                            finalResults.add(resultWithQuery);
-                        }
-                    }
-                } else {
-                    // For other experiment types, use the original format
-                    queryResults.put(PAIRWISE_FIELD_NAME_QUERY_TEXT, queryText);
-                    finalResults.add(queryResults);
-                }
-
+                queryResults.put(PAIRWISE_FIELD_NAME_QUERY_TEXT, queryText);
+                finalResults.add(queryResults);
                 if (pendingQueries.decrementAndGet() == 0) {
                     updateFinalExperiment(experimentId, request, finalResults, judgmentList);
                 }
