--- src/test/resources/searchconfig/CreateSearchConfigurationSimpleMatch.json
@@ -1,5 +1,0 @@
-{
-  "name": "simple_match_query",
-  "index": "{{index_name}}",
-  "query": "{\"query\":{\"match\":{\"title\":\"%SearchText%\"}}}"
-}
